invoice: fix monthly timestamp generation
The code currently generates the final/monthly invoice for timestamp - 1
day - 1 month, which effectively is two months ago when being called at
the first day of the month.

There is a follow-up issue, that the code generates the invoice subject
for two-month-ago time-range. Next it fetches the list of users, which
needs an invoice for the time range. But then it gets the invoice
content based on the correct time range (i.e. the previous month). This
can result in fatal failures when the code tries to generate an invoice
for a user, which bought something two months ago, but did not buy
anything in the previous month.