web: add special handling for missing product
Add a special page handler when opening the product detail page for
an EAN unknown to the system. This allows opening the detail page
for an arbitrary EAN and follow the flow to either create the product
first or restock it directly. It's very useful when being combined
with a barcode scanner.