web: prepare new restock feature
It's often necessary to restock multiple products at the same
time. This is currently requires going to the detail page of
each product to access the right form, which is quite cumbersome.

This prepares a new page to restock multiple products at once,
which hopefully makes this a less tedious job.